@@ -1,16 +1,24 @@
 {
+  "name": "realworld-vue",
+  "version": "0.1.0",
+  "description": "TodoMVC for the RealWorld™",
   "author": "Emmanuel Vilsbol <emmanuel@vilsbol.com>",
+  "scripts": {
+    "serve": "cross-env BABEL_ENV=dev vue-cli-service serve",
+    "build": "cross-env BABEL_ENV=dev vue-cli-service build",
+    "lint": "vue-cli-service lint",
+    "test": "cross-env BABEL_ENV=test jest --coverage"
+  },
   "dependencies": {
     "axios": "^0.19.0",
     "date-fns": "^1.30.1",
     "marked": "^0.7.0",
     "register-service-worker": "^1.6.2",
-    "vue": "^2.6.10",
+    "vue": "^3.0.0-beta.1",
     "vue-axios": "^2.1.4",
     "vue-router": "^3.1.3",
-    "vuex": "^3.1.1"
+    "vuex": "^4.0.0-alpha.1"
   },
-  "description": "TodoMVC for the RealWorld™",
   "devDependencies": {
     "@babel/core": "^7.6.2",
     "@babel/preset-env": "^7.6.2",
@@ -19,15 +27,18 @@
     "@vue/cli-plugin-pwa": "^3.10.0",
     "@vue/cli-plugin-unit-jest": "^3.10.0",
     "@vue/cli-service": "^3.10.0",
+    "@vue/compiler-sfc": "^3.0.0-beta.1",
     "@vue/eslint-config-prettier": "^4.0.1",
-    "@vue/test-utils": "^1.0.0-beta.29",
+    "@vue/test-utils": "^2.0.0-alpha.1",
     "babel-core": "7.0.0-bridge.0",
     "babel-jest": "^24.9.0",
     "cross-env": "^6.0.3",
+    "eslint-plugin-vue": "^7.0.0-alpha.0",
     "lint-staged": "^8.2.1",
     "node-sass": "^4.12.0",
     "nyc": "^14.1.1",
     "sass-loader": "^7.1.0",
+    "vue-cli-plugin-vue-next": "~0.1.4",
     "vue-template-compiler": "^2.6.10"
   },
   "gitHooks": {
@@ -42,13 +53,5 @@
       "vue-cli-service lint",
       "git add"
     ]
-  },
-  "name": "realworld-vue",
-  "scripts": {
-    "build": "cross-env BABEL_ENV=dev vue-cli-service build",
-    "lint": "vue-cli-service lint",
-    "serve": "cross-env BABEL_ENV=dev vue-cli-service serve",
-    "test": "cross-env BABEL_ENV=test jest --coverage"
-  },
-  "version": "0.1.0"
+  }
 }
