@@ -1,4 +1,4 @@
-import Vue from "vue";
+import Vue, { createApp } from "vue";
 import App from "./App.vue";
 import router from "./router";
 import store from "./store";
@@ -9,7 +9,6 @@ import ApiService from "./common/api.service";
 import DateFilter from "./common/date.filter";
 import ErrorFilter from "./common/error.filter";
 
-Vue.config.productionTip = false;
 Vue.filter("date", DateFilter);
 Vue.filter("error", ErrorFilter);
 
@@ -20,8 +19,13 @@ router.beforeEach((to, from, next) =>
   Promise.all([store.dispatch(CHECK_AUTH)]).then(next)
 );
 
-new Vue({
-  router,
-  store,
-  render: h => h(App)
-}).$mount("#app");
+// v2 syntax
+// new Vue({
+//   router,
+//   store,
+//   render: h => h(App)
+// }).$mount("#app");
+createApp(App)
+  .use(router)
+  .use(store)
+  .mount("#app");
