@@ -1,4 +1,3 @@
-import Vue from "vue";
 import Vuex from "vuex";
 
 import home from "./home.module";
@@ -6,9 +5,13 @@ import auth from "./auth.module";
 import article from "./article.module";
 import profile from "./profile.module";
 
-Vue.use(Vuex);
+// v2 syntax
+// Vue.use(Vuex);
+// export default new Vuex.Store({
+//   modules: { ... }
+// });
 
-export default new Vuex.Store({
+export default Vuex.createStore({
   modules: {
     home,
     auth,
